refactor(core): extract pod/health.py — pure code move
Track A Sprint 1 Step 1 (pod_health). Splits the former
src/winpodx/core/pod.py into a package:

- pod/health.py: check_rdp_port, recover_rdp_if_needed (and the
  module-level subprocess/time/socket/re imports they rely on).
- pod/backend.py: PodState, PodStatus, get_backend, pod_status.
- pod/__init__.py: re-exports the full public surface so existing
  'from winpodx.core.pod import X' callers keep working without churn.
  start_pod and stop_pod stay here for now; they move to pod/lifecycle.py
  in the next commit.

Behaviour is unchanged. Tests that reach in via module-attribute patches
(monkeypatch.setattr('winpodx.core.pod.subprocess.run', ...) and friends)
are updated to point at the actual implementation module — these patches
are the rare case where shimming the public surface isn't enough, since
they target internal globals consumed by the patched function's own
module scope.
