+++ src/winpodx/core/pod/backend.py
@@ -1,0 +1,79 @@
+"""Pod state types, backend factory, and status query."""
+
+from __future__ import annotations
+
+import logging
+from dataclasses import dataclass
+from enum import Enum
+
+from winpodx.backend.base import Backend
+from winpodx.core.config import Config
+from winpodx.core.pod.health import check_rdp_port
+
+log = logging.getLogger(__name__)
+
+
+class PodState(Enum):
+    STOPPED = "stopped"
+    STARTING = "starting"
+    RUNNING = "running"
+    PAUSED = "paused"
+    ERROR = "error"
+
+
+@dataclass
+class PodStatus:
+    state: PodState
+    ip: str = ""
+    uptime: str = ""
+    cpu_usage: str = ""
+    memory_usage: str = ""
+    error: str = ""
+
+
+def get_backend(cfg: Config) -> Backend:
+    """Instantiate the appropriate backend based on config."""
+    name = cfg.pod.backend
+    if name == "docker":
+        from winpodx.backend.docker import DockerBackend
+
+        return DockerBackend(cfg)
+    elif name == "podman":
+        from winpodx.backend.podman import PodmanBackend
+
+        return PodmanBackend(cfg)
+    elif name == "libvirt":
+        from winpodx.backend.libvirt import LibvirtBackend
+
+        return LibvirtBackend(cfg)
+    elif name == "manual":
+        from winpodx.backend.manual import ManualBackend
+
+        return ManualBackend(cfg)
+    else:
+        raise ValueError(f"Unknown backend: {name}")
+
+
+def pod_status(cfg: Config) -> PodStatus:
+    """Query the current pod status."""
+    backend = get_backend(cfg)
+    try:
+        running = backend.is_running()
+    except Exception as e:
+        log.error("Failed to query pod status: %s", e)
+        return PodStatus(state=PodState.ERROR, error=str(e))
+
+    if not running:
+        return PodStatus(state=PodState.STOPPED)
+
+    try:
+        if backend.is_paused():
+            return PodStatus(state=PodState.PAUSED, ip=cfg.rdp.ip)
+    except Exception as e:  # pragma: no cover - defensive
+        log.debug("is_paused probe failed: %s", e)
+
+    rdp_ok = check_rdp_port(cfg.rdp.ip, cfg.rdp.port)
+    return PodStatus(
+        state=PodState.RUNNING if rdp_ok else PodState.STARTING,
+        ip=cfg.rdp.ip,
+    )
