+++ src/winpodx/core/pod/health.py
@@ -1,4 +1,4 @@
-"""Pod lifecycle management."""
+"""RDP-port liveness probes and pod recovery helpers."""
 
 from __future__ import annotations
 
@@ -7,57 +7,13 @@ import re
 import socket
 import subprocess
 import time
-from dataclasses import dataclass
-from enum import Enum
 
-from winpodx.backend.base import Backend
 from winpodx.core.config import Config
 
 log = logging.getLogger(__name__)
 
 # Container name guard reused for the recovery-path subprocess args.
 _CONTAINER_NAME_RE = re.compile(r"^[a-zA-Z0-9][a-zA-Z0-9._-]{0,127}$")
-
-
-class PodState(Enum):
-    STOPPED = "stopped"
-    STARTING = "starting"
-    RUNNING = "running"
-    PAUSED = "paused"
-    ERROR = "error"
-
-
-@dataclass
-class PodStatus:
-    state: PodState
-    ip: str = ""
-    uptime: str = ""
-    cpu_usage: str = ""
-    memory_usage: str = ""
-    error: str = ""
-
-
-def get_backend(cfg: Config) -> Backend:
-    """Instantiate the appropriate backend based on config."""
-    name = cfg.pod.backend
-    if name == "docker":
-        from winpodx.backend.docker import DockerBackend
-
-        return DockerBackend(cfg)
-    elif name == "podman":
-        from winpodx.backend.podman import PodmanBackend
-
-        return PodmanBackend(cfg)
-    elif name == "libvirt":
-        from winpodx.backend.libvirt import LibvirtBackend
-
-        return LibvirtBackend(cfg)
-    elif name == "manual":
-        from winpodx.backend.manual import ManualBackend
-
-        return ManualBackend(cfg)
-    else:
-        raise ValueError(f"Unknown backend: {name}")
 
 
 def check_rdp_port(ip: str, port: int, timeout: float = 5.0) -> bool:
@@ -67,37 +23,6 @@ def check_rdp_port(ip: str, port: int, timeout: float = 5.0) -> bool:
             return True
     except (ConnectionRefusedError, TimeoutError, OSError):
         return False
-
-
-def start_pod(cfg: Config) -> PodStatus:
-    """Start the Windows pod and wait up to boot_timeout for RDP readiness."""
-    backend = get_backend(cfg)
-    try:
-        backend.start()
-    except Exception as e:
-        log.error("Failed to start pod: %s", e)
-        return PodStatus(state=PodState.ERROR, error=str(e))
-
-    try:
-        ready = backend.wait_for_ready(timeout=cfg.pod.boot_timeout)
-    except Exception as e:
-        log.error("wait_for_ready failed: %s", e)
-        return PodStatus(state=PodState.ERROR, error=str(e))
-
-    if ready:
-        return PodStatus(state=PodState.RUNNING, ip=cfg.rdp.ip)
-    return PodStatus(state=PodState.STARTING, ip=cfg.rdp.ip)
-
-
-def stop_pod(cfg: Config) -> PodStatus:
-    """Stop the Windows pod."""
-    backend = get_backend(cfg)
-    try:
-        backend.stop()
-    except Exception as e:
-        log.error("Failed to stop pod: %s", e)
-        return PodStatus(state=PodState.ERROR, error=str(e))
-    return PodStatus(state=PodState.STOPPED)
 
 
 def recover_rdp_if_needed(cfg: Config, *, max_attempts: int = 3) -> bool:
@@ -163,28 +88,3 @@ def recover_rdp_if_needed(cfg: Config, *, max_attempts: int = 3) -> bool:
 
     log.warning("RDP recovery exhausted %d attempts; RDP still down.", max_attempts)
     return False
-
-
-def pod_status(cfg: Config) -> PodStatus:
-    """Query the current pod status."""
-    backend = get_backend(cfg)
-    try:
-        running = backend.is_running()
-    except Exception as e:
-        log.error("Failed to query pod status: %s", e)
-        return PodStatus(state=PodState.ERROR, error=str(e))
-
-    if not running:
-        return PodStatus(state=PodState.STOPPED)
-
-    try:
-        if backend.is_paused():
-            return PodStatus(state=PodState.PAUSED, ip=cfg.rdp.ip)
-    except Exception as e:  # pragma: no cover - defensive
-        log.debug("is_paused probe failed: %s", e)
-
-    rdp_ok = check_rdp_port(cfg.rdp.ip, cfg.rdp.port)
-    return PodStatus(
-        state=PodState.RUNNING if rdp_ok else PodState.STARTING,
-        ip=cfg.rdp.ip,
-    )
